feature: Models - change id placement
