+++ apps/ai-dial-admin/src/app/[lang]/models/[id]/page.tsx
@@ -1,16 +1,14 @@
 import { cookies, headers } from 'next/headers';
 import { notFound } from 'next/navigation';
 
-import { getCoreModel, removeModel, updateCoreModel, updateModel } from '@/src/app/[lang]/models/actions';
 import { interceptorsApi, modelsApi, rolesApi } from '@/src/app/api/api';
-import EntityView from '@/src/components/EntityView/View/EntityView';
+import View from '@/src/components/Models/View/View';
 import { DEFAULT_ETAG } from '@/src/constants/api-headers';
 import { SaveValidationContextProvider } from '@/src/context/SaveValidationContext';
 import { DialInterceptor } from '@/src/models/dial/interceptor';
 import { DialModel } from '@/src/models/dial/model';
 import { DialRole } from '@/src/models/dial/role';
 import { errorObjLog } from '@/src/server/logger';
-import { ApplicationRoute } from '@/src/types/routes';
 import { getUserToken } from '@/src/utils/auth/auth-request';
 import { filterDisplayNamesWithVersions } from '@/src/utils/entities/filter-names';
 import { getIsEnableAuthToggle } from '@/src/utils/env/get-auth-toggle';
@@ -45,18 +43,7 @@ export default async function Page(params: { params: Promise<{ id: string }> }) 
 
   return (
     <SaveValidationContextProvider>
-      <EntityView
-        view={ApplicationRoute.Models}
-        names={names}
-        etag={etag}
-        roles={roles}
-        interceptors={interceptors}
-        originalEntity={model}
-        removeEntity={removeModel}
-        updateEntity={updateModel}
-        getCoreEntity={getCoreModel}
-        updateCoreEntity={updateCoreModel}
-      />
+      <View names={names} etag={etag} roles={roles} interceptors={interceptors} originalModel={model} />
     </SaveValidationContextProvider>
   );
 }
